#193481: Use NIO's FileLock to take ownership of lock file

+++ o.n.bootstrap/src/org/netbeans/CLIHandler.java
@@ -44,22 +44,25 @@
 
 package org.netbeans;
 
+import java.io.Closeable;
 import java.io.DataInputStream;
 import java.io.DataOutputStream;
 import java.io.EOFException;
 import java.io.File;
 import java.io.FileInputStream;
-import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InterruptedIOException;
 import java.io.OutputStream;
 import java.io.PrintWriter;
+import java.io.RandomAccessFile;
 import java.net.InetAddress;
 import java.net.ServerSocket;
 import java.net.Socket;
 import java.net.SocketException;
 import java.net.UnknownHostException;
+import java.nio.channels.FileLock;
+import java.nio.channels.OverlappingFileLockException;
 import java.security.NoSuchAlgorithmException;
 import java.security.SecureRandom;
 import java.util.ArrayList;
@@ -129,7 +132,7 @@ public abstract class CLIHandler extends Object {
     /** Testing output of the threads.
      */
     private static final Logger OUTPUT = Logger.getLogger(CLIHandler.class.getName());
-    
+
     private int when;
     
     /**
@@ -327,6 +330,16 @@ public abstract class CLIHandler extends Object {
         }
     }
     
+    private static FileLock tryLock(RandomAccessFile raf) throws IOException {
+        try {
+            return raf.getChannel().tryLock();
+        } catch (OverlappingFileLockException ex) {
+            // happens in CLIHandlerTest as it simulates running multiple
+            // instances of the application in the same VM
+            return null;
+        }
+    }
+    
     
     /** Initializes the system by creating lock file.
      *
@@ -504,8 +517,18 @@ public abstract class CLIHandler extends Object {
         
         for (int i = 0; i < 5; i++) {
             // try few times to succeed
+            RandomAccessFile raf = null;
+            FileLock lock = null;
             try {
-                if (lockFile.exists()) {
+                NO_LOCK: if (lockFile.isFile()) {
+                    
+                    raf = new RandomAccessFile(lockFile, "rw");
+                    lock = tryLock(raf);
+                    if (lock != null && lock.isValid()) {
+                        // not locked by anyone,
+                        break NO_LOCK;
+                    }
+                    
                     enterState(5, block);
                     throw new IOException("EXISTS"); // NOI18N
                 }
@@ -514,8 +537,15 @@ public abstract class CLIHandler extends Object {
                     return new Status(2);
                 }
                 
-                lockFile.getParentFile().mkdirs();
-                lockFile.createNewFile();
+                if (lock == null) {
+                    lockFile.getParentFile().mkdirs();
+                    try {
+                        raf = new RandomAccessFile(lockFile, "rw");
+                        lock = tryLock(raf);
+                    } catch (IOException ex) {
+                        throw ex;
+                    }
+                }
                 lockFile.deleteOnExit();
                 secureAccess(lockFile);
                 
@@ -523,13 +553,14 @@ public abstract class CLIHandler extends Object {
                 
                 final byte[] arr = new byte[KEY_LENGTH];
                 new Random().nextBytes(arr);
+
                 
-                server = new Server(arr, block, handlers, failOnUnknownOptions);
-                
-                final DataOutputStream os = new DataOutputStream(new FileOutputStream(lockFile));
+                final RandomAccessFile os = raf;
+                raf.seek(0L);
+                server = new Server(os, arr, block, handlers, failOnUnknownOptions);
                 int p = server.getLocalPort();
                 os.writeInt(p);
-                os.flush();
+                os.getChannel().force(true);
                 
                 enterState(20, block);
                 
@@ -559,7 +590,7 @@ public abstract class CLIHandler extends Object {
 
                         try {
                             os.write(arr);
-                            os.flush();
+                            os.getChannel().force(true);
 
                             enterState(27,block);
                             // if this turns to be slow due to lookup of getLocalHost
@@ -582,7 +613,7 @@ public abstract class CLIHandler extends Object {
                             ex.printStackTrace();
                         }
                         try {
-                            os.close();
+                            os.getChannel().force(true);
                         } catch (IOException ex) {
                             // ignore
                         }
@@ -810,6 +841,11 @@ public abstract class CLIHandler extends Object {
                         }
                     }
                     
+                    if (lock == null) {
+                        // process exists (we cannot lock the file), but does not respond
+                        return new Status (Status.CANNOT_CONNECT);
+                    }
+                    
                     if (cleanLockFile || isSameHost) {
                         // remove the file and try once more
                         lockFile.delete();
@@ -937,6 +973,7 @@ public abstract class CLIHandler extends Object {
     /** Server that creates local socket and communicates with it.
      */
     private static final class Server extends Thread {
+        private Closeable unlock;
         private byte[] key;
         private ServerSocket socket;
         private Integer block;
@@ -949,8 +986,9 @@ public abstract class CLIHandler extends Object {
         /** by default wait 100ms before sending a REPLY_FAIL message */
         private static long failDelay = 100;
         
-        public Server(byte[] key, Integer block, Collection<? extends CLIHandler> handlers, boolean failOnUnknownOptions) throws IOException {
+        public Server(Closeable lock, byte[] key, Integer block, Collection<? extends CLIHandler> handlers, boolean failOnUnknownOptions) throws IOException {
             super("CLI Requests Server"); // NOI18N
+            this.unlock = lock;
             this.key = key;
             this.setDaemon(true);
             this.block = block;
@@ -1030,6 +1068,13 @@ public abstract class CLIHandler extends Object {
         }
 
         final void stopServer () {
+            if (unlock != null) {
+                try {
+                    unlock.close();
+                } catch (IOException ex) {
+                    OUTPUT.log(Level.WARNING, "Cannot unlock {0}", ex.getMessage());
+                }
+            }
             socket = null;
             // interrupts the listening server
             interrupt();
